kubernetes-csi: enable Docker
We must explicitly call runner.sh and that in turn needs a privileged
security context.

The resource allocation is the same as in
config/jobs/kubernetes/sig-testing/verify.yaml, which served as
example for a job using Docker.

preset-cadvisor-docker-credential isn't needed, that was a
cut-and-paste error when setting up the job.