Add more feedback for users of /override
Two of the most common issues we see after rolling out the override
plugin are that admins will not specify a context to override, or they
will specify the "common name" of the context that they would use to
trigger the job in the first place. We should give them feedback for
these cases.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>