Update sig-windows-config.yaml
Increases staging job frequency to 3h. 

With the merging of: https://github.com/kubernetes/kubernetes/pull/71950 the two tests mentioned here: https://github.com/kubernetes/kubernetes/issues/71949 should be ok. Added both of them to the inclusion list of the staging job. Once we have a couple of successful runs, we move them to the main job and 1.14 job.