Update sig-windows-config.yaml
Removed GMSA from staging job. Reason: GMSA test cannot run on staging job ATM as the cluster is not configured properly to accommodate it.

Added ginkgo.flakeAttempts=2 and removed the two skipped tests from main job. It seems this combination behave well on the staging job, we're adding it to the master job with the intention to move to the release job (1.14) once we have more data.