Allow the commenter to iterate randomly through results
When the commenter has a low ceiling compared to the total result set,
it is very common for it to target a small subset of alphabetically
first results to comment on, on every iteration. A new flag allows the
commenter to be configured to choose issues at random from the result
pool.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>