kubernetes-csi: do not test deployment changes
It turned out that the case that these jobs were meant to catch ("does
the deployment work as specified without changing images") is too
strict. The plan is to allow bumping the hostpath image version before
actually publishing that image, to ensure that the tagged release has
the right image version. This will break the deployment temporarily on
the branch.