kubernetes-csi: split up jobs into unit/alpha/non-alpha
This is meant to help developers who look at tests results understand
more quickly what the root cause of a failure might be because they
don't need to look at the individual test name.

Splitting up the pre-merge jobs also speeds up providing feedback on a
PR, because jobs complete faster.

Naming of jobs is a bit more consistent now, with "kubernetes-csi"
being used also for pre-merge jobs. csi-driver-host-path
now has:

pull-kubernetes-csi-csi-driver-host-path-1-13-on-kubernetes-1-13
pull-kubernetes-csi-csi-driver-host-path-1-13-on-kubernetes-master
pull-kubernetes-csi-csi-driver-host-path-1-14-on-kubernetes-1-14
pull-kubernetes-csi-csi-driver-host-path-1-14-on-kubernetes-master
pull-kubernetes-csi-csi-driver-host-path-alpha-1-13-on-kubernetes-1-13 (*)
pull-kubernetes-csi-csi-driver-host-path-alpha-1-14-on-kubernetes-1-14 (*)
pull-kubernetes-csi-csi-driver-host-path-unit (*)

ci-kubernetes-csi-1-13-on-kubernetes-1-13
ci-kubernetes-csi-1-13-on-kubernetes-1-14
ci-kubernetes-csi-1-13-on-kubernetes-master
ci-kubernetes-csi-alpha-1-13-on-kubernetes-1-13 (*)
ci-kubernetes-csi-canary-on-kubernetes-1-13
ci-kubernetes-csi-canary-on-kubernetes-1-14
ci-kubernetes-csi-canary-on-kubernetes-master
ci-kubernetes-csi-alpha-canary-on-kubernetes-master (*)

Jobs marked (*) with are new, other others have been changed to not
run alpha tests.

The "liveness-probe" pre-merge tests were limited to run only on one
Kubernetes version because there is no real dependency on Kubernetes
APIs.