Fix poll optimization in SecretAgent
Properly reset the counter for skips (=not loading a secret because the
underlying file was not observably changed) when the secret is reloaded.

The watch mechanism is supposed to reload a secret even after no changes
were observed for X iterations. The counter for skips was never zeroed
so after 10 minutes the Agent reloaded the secrets each second.