Re-enable prometheus in GCE 5K Performance tests.
We've ruled out that the prometheus was a cause of the regressions a long time ago. Actually it helps a lot when debugging.

The positive side effect of this change will be a mitigation of the faster MIG startup problem, in the end it will add an O(a few minutes) delay before running the tests.

Ref. https://github.com/kubernetes/kubernetes/issues/76579