Add ChangedFiles to ProwJobSpec via Pull struct
To allow the downward api to provide the list of files changed by a PR,
added a new ChangedFiles field to Pull struct. This allows the list of
files changed by a PR to be included in Refs of a ProwJobSpec, later
supplied to a Prow job (either PreSubmit or PostSubmit).

Signed-off-by: Arash Deshmeh <adeshmeh@ca.ibm.com>