Use cross-build for GKE tests
GKE now expects Windows artifacts in a release. Windows artifacts
are not included in bazel builds, because we don't cross-build with
bazel. This caused a number of test failures last week and we
implemented an internal hotfix to help, but the correct fix is to test
builds that include the required artifacts.