Made kubeflow-pipeline-sample-test mandatory
This test is important and the consensus is that it's desirable to make this test mandatory for PRs.
The e2e test runs for 45 minutes and the samples test runs for 60 minutes, so making the test mandatory adds no more than 15 minutes of wait time. In reality this will probably save lots of human time and prevent broken releases.