Re-run config-forker for v1.15
I did this mostly as a test to see if the latest changes make
it idempotent vs. the manual changes that have already been
made to the v1.15 config. It seems like there have been a few
changes to master during this time, and given that we're still
in code freeze I think they're all still relevant to the 1.15
jobs.  How do we feel about a refresh?

Oh right, also ran ./hack/update-config.sh after that