Record ProwJob state transitions in metrics
When calculating success rates, for instance, using the current set of
metrics that expose the distribution of ProwJob custom resources in the
cluster's data store, a dip in the success rate may happen for one of
two reasons:

 - new observed ProwJob CRs are failed
 - previously observed successful ProwJob CRs were garbage collected

This property means that the current set of metrics, while useful to
show current footprint on the APIServer, are not useful for higher order
calculations. This change introduces counters for state transitions by
introducing hysteresis to the metrics recorder in order to publish
metrics that are appropriate for that purpose.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>