Remove CAPV periodic test
The CAPV e2e tests now leverage KIND for cluster deployment, and
therefor no longer need to build a new OVA as part of the periodic Prow
Job. Once we remove that step from the e2e test, the periodic job is
identical to the presubmit job. Until they are different (which may
happen again in the future) there is no point in tying up infrastructure
for that test. So we remove the periodic entirely.