Update kind conformance test to build go-runner
A PR in k/k is adding a test binary into the conformance
image in an effort to expand capabilities and reliability.

Due to lack of coordination between repos, this test-infra
script currently fails on the proposed PR because it doesn't
know about the new binary.

This change tries to build that new component if it is there
but will skip it otherwise. This way the timing
between the PRs is more forgiving.

A follow-up should build it like the other components, causing
a failure if it is not built correctly every time.