Only test/check go source if they've changed
This patch modifies the tests that do go fmt|vet|lint to only run if the
go sources in `pkg` or `cmd` have changed. Additionally, we only need to
see if the build actually builds if the go sources have changed, or
go.mod, which indicates a change in dependencies.

This means that a simple doc change won't re-run all the Go tests.