checkconfig: add optional check to verify OWNERS presence
When a repo enables an approve plugin but does not contain OWNERS
file, it does not make sense and the situation pollutes the logs.

The file presence is checked using `RepositoryClient.GetFile()` method
which transfers the content that we do not actually use, but given
OWNERS files are generally very small, I think it is not worth
implementing something like a `HasFile` with a HEAD request.

Because this check is not local and cheap, I did not make it run by
default and introduced a new option to enable "expensive" checks.