Tide: Only retest presubmits that don't have an up-to-date result
Currently, when Tide merges serially instead of in batches and the PR is
missing at least one Presubmit, Tide will create new jobs for _all_
Presubmits the PR requires. This has some drawbacks:

* We waste capacity of our build cluster
* Those newly-created jobs for Presubmits that were already passing will
  report to GitHub, even thought Tide doesn't care about their result
  because there is already a successful job. This is confusing for
  users, because it looks like Tide merges PRs even thought some of the
  jobs are still pending or even failing.

This PR changes tide to keep track of the Presubmits that are actually
missing and use that info when doing serial retesting.