Add a specific base image for boskosctl
Users of boskosctl inside of a container will require `jq` as well as
`bash` in order to process the resource records that `boskosctl` prints
when leases are appropriately made. The `alpine-bash` image may remain
useful as a base image for other derivatives, so no change is made
there, but a new image is specifically made for `boskosctl` so adding
more packages to this image is simpler.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>