Record GitHub request durations in a histogram
When adding labels to a Prometheus metric, the set of label values
should be bounded in order to bound the memory and CPU footprint that
the metric has on the process exposing it as well as the Prometheus
server scraping it and serving requests for it. Durations of requests
are better exposed as a histogram with the same set of identifying
labels as the counters, instead of as a label.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>