swap periodic bazel jobs for postsubmits in sig-release boards
There are postsubmit and periodic variants of bazel test and build
jobs on sig-release dashboards. The postsubmits are currently in the
-blocking and -all boards, and the periodics in -all.

Jobs on release-blocking boards are expected to run at least every
3 hours. The postsubmits fail to meet this expectation since they
tend to go quiet on the weekend when no commits are coming in.

Reconcile to the following state for all but master
- postsubmit bazel-foo goes to sig-release-1.y-all
- periodic bazel-foo goes to sig-release-1.y-blocking, -all, google-unit

For master:
- postsubmit bazel-foo goes to google-unit
- periodic bazel-foo goes to sig-release-master-blocking, -all,
google-unit
- bazel-foo-canary untouched, just field reordering

I also adjusted the ordering of `name:` and `annotation:` fields for
human readability