Remove testgrid annotations from forked presubmits.
Presubmits do not get renamed, which results in their forks being
duplicates. In some cases this is a detectable error, but in other cases
the configurations of each version will be merged (with properties
being overwritten with arbitrary precedence).

To avoid the problem, just drop the testgrid annotations from the
duplicate jobs.