bugzilla: handle comparing statuses with resolutions
Bugzilla records in final state also have a "Resolution" sub-state,
which is usually used to store info about how the bug was resolved
(fixed, refused, etc.). This is essentially a part of the status and so
it should be possible to make decisions based on it. This commit allows
to optionally specify valid statuses including a resolution in config:

```
- CLOSED (ERRATA)
```

In this case, only `CLOSED` bugs with `ERRATA` resolution would be
considered valid. It is still possible to ignore resolutions:

```
- CLOSED
```

In this case, just the status will be used for validity check.