Be more lenient with errors in heartbeats
When the boskosctl command fails to send one heartbeat, it is not a
catastrophic error as we send heartbeats much more frequently than
strictly necessary to retain our lease. This allows us to keep an error
budget and exit only when exceeding it, which is valuable as transitive
network issues should not crash the heartbeat command.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>