Run pull-kubernetes-e2e-gce-alpha-features when feature related files are touched
Currently this job is run only on demand as folks do not know that this
job exists. So let's try to run this job when any of the files related
to feature flags are touched. This will ensure folks who are adding
alpha features or promoting stuff end up running this job when they
touch those files. Currently we end up triaging the corresponding
periodic CI job too late in the cycle. We should warn folks early by
automatically running this presubmit job instead.

Change-Id: Icfe7f0ba2623be0a803a924b570c9d8d35cb6ed8