bugzilla: introduce states to eventually replace statuses
This commit does not yet change the external functionality of the
plugin. The purpose of this commit is to allow backward compatibility
for the existing config files using `status` fields by mirroring them
into `state` fields at runtime. There is some tricky logic for handling
the presence of both in the same config. The bugzilla plugin still uses
"status" fields only, so this commit only prepares the stage for
switching it to status fields.