Consider a corner case in ExpandAliases
The corner case: A team has no members (yet) in OWNERS_ALIASES:
Although it is arguable why it introduces a team without
members in the first place, it is still a valid yaml file.

OWNERS
```
reviewers:
  - team-name
```

OWNERS_ALIASES
```
aliases:
  team-name:
```

When expanding, we replace the team by empty set, instead of
ignoring it. Otherwise, the team would be left there as a username,
which leads to a violation of `verify-owners`, saying `team-name`
is not a member of ...

Also added the test for this case.