kubernetes-csi: avoid overriding Kubernetes version for single-Kubernetes repos
The original intention was to let repos which only need to be tested
against a single Kubernetes release choose that release
themselves. The comment for `single_kubernetes_repos` still describes
that behavior.

https://github.com/kubernetes/test-infra/pull/13434 changed that so
that 1.15 is used for all of these repos because of some issues with
external-snapshotter.

external-snapshotter now has 1.15 in its own prow.sh, so we don't need
the override anymore in the test job and can revert to the original
behavior.