prow/cluster/starter: plank needs to patch prowjobs.
I just stood up prow using the "tackle" utility.  The plank Role
created was not sufficient.  The "echo-test" job was not running.
This was because plank was emitting errors in its log about not being
able to patch the prowjob.  Manually adding this to the Role fixed the
problem for me.