Print the SHA for each repository
Currently when we have a CI job with many repos, we don't really print
the SHA for each repository AFAICT. Here's an attempt to do just that.

Example:
https://storage.googleapis.com/kubernetes-jenkins/logs/ci-cluster-api-provider-gcp-make-conformance/1178735268761440257/build-log.txt

The started.json has `a87c1b2d720fce` which is the SHA for k/k and
i can't find which SHA was used when checking out
sigs.k8s.io/cluster-api-provider-gcp. the build.log also has the SHA for
k/k but not the repo i was looking for.