use kubekins-e2e image for build and release
This patch replaces the use of the "csi-ci" image, as it is not well
maintained and the latest version of it fails to start up DinD. Rather
than fix/maintain that, use kubekins-e2e. While the original motivation
for "csi-ci" was to speed up the builds by caching pre-compiled packages
and pre-downloading the go module dependencies, the recent usage of
GOPROXY has made most of the time spent downloading deps a non-issue.