images/builder: Optionally force logging to stdout with `LOG_TO_STDOUT`
Here we add a check for an env var called `LOG_TO_STDOUT`.
If it is set to `y|Y`, we will force logs to stdout instead of a
specified log directory.

This is useful for CI instances where a single variant is run.
The current behavior writes to a build log which you need to dig for a
little, as opposed to being able to view via prow.k8s.io/view.

This also fixes some shellcheck errors:
- [[ ]] undefined in POSIX: https://github.com/koalaman/shellcheck/wiki/SC2039
- Use -n instead of ! -z: https://github.com/koalaman/shellcheck/wiki/SC2236

Signed-off-by: Stephen Augustus <saugustus@vmware.com>