conformance: Remove ci-kubernetes-gce-conformance-stable-* jobs
As these are periodic jobs targeted against the release/* version
markers, which represent released versions of Kubernetes, we don't gain
much benefit from testing them.

We already have the set of ci-kubernetes-gce-conformance-latest* jobs,
which fork when we generate new release branch jobs, and are included on
all release-blocking dashboards.

These tests are the last tests labeled as
"manual-release-bump-required".

Signed-off-by: Stephen Augustus <saugustus@vmware.com>