Use kubekins-e2e image everywhere possible
This patch replaces the CPI "CI" image with the kubekins-e2e image. The
Docker-in-Docker functionality in the CI image was no longer working,
and wit the improvements to GOPROXY support, the CI Image is no longer
strictly necessary. This way we get a constantly supported base image
that is already maintained.

The exception here is that the periodic conformance tests still use the
CI image as those have additional dependencies on sk8.