ci-k8sio-backup: remove job
The problem here is that the backup.sh script itself invokes a Docker
image *and* that this job is meant to run in the trusted
cluster (because of the credential secrets involved). It's not a good
idea to run containers in the trusted cluster, and so this should run
in a different K8s cluster, outside of Prow.

This is a TODO item for this job anyway (see the diff), so it looks like
it will have to be done as part of the initial implementation.

This commit manually reverts the following PRs:

- #14904
- #15185
- #15197