Support nested metadata in the Spyglass metadata lens.
The metadata lens now supports nested JSON objects. For nested JSON
objects, the name of the objects is prepended to the name of the field
for visualization. For example, if you had metadata that looked like
this:

{
  "field1": "value1",
  "nested-parent": {
    "nested-field1": "nested-value1",
    "nested-field2": "nested-value2"
  }
}

would produce a table that looks like this:

field1: value1
nested-parent.nested-field1: nested-value1
nested-parent.nested-field2: nested-value2

Otherwise, the current behavior of the metadata lens has been
maintained.