Pass issue_comment events to needs-rebase plugin
The needs-rebase plugin operates in three ways:

* evaluate merge-ability when the PR changes code
* reevaluate merge-ability when a PR gets a comment
* sweep open PRs on a configured interval to reevaluate

This pull (correctly) sends issue comment events to the plugin so
it can use them. That's more efficient than trying to re-sync all
PRs more frequently. A re-sync/re-evaluation is necessary in the
first place since the plugin does not map push events from the
branch to open PRs for it.