add pull-k8sio-backup presubmit check
This checks the backup_tools scripts by testing them. Specifically, it
runs the backup_test.sh script, which performs the backup from

    us.gcr.io/k8s-gcr-backup-test-prod

    to

    us.gcr.io/k8s-gcr-backup-test-prod-bak

for a set of chosen images. Both this test and the real backup logic in
backup_prod.sh share the same "build_gcrane" and "copy_with_date"
functions --- the main difference is the repositories where the backups
happen and the images that are backed up.