Increase CAPA e2e conformance timeout
Increase the timeout for
ci-cluster-api-provider-aws-e2e-conformance-release-0-4 to 3 hours (the
default is 2). This job is consistently timing out, and it's unclear if
it needs more time, or if there is some other issue. If it does not
finish in 3 hours, then we know there is some other issue we need to
investigate.

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>