generate_tests: Use "latest" markers instead of "k8s-*" pseudo-versions
The "k8s-beta" version marker leads to confusion as sometimes "beta"
should refer to the current pre-release branch and sometimes it should
point to master. There is no consistent guidance or follow-through on
how or when these markers need to be changed, which can lead to a slew
of tests that don't test what we think they do.

As a first step, we're setting test_config.yaml to use the "latest**"
version markers, which are pretty clear at a glance which version of
Kubernetes will be tested against:

- "dev" --> "latest" --> latest CI build of pre-1.18.0 (master)
- "beta" --> "latest-1.17" --> latest CI build of release-1.17
- "stable1" --> "latest-1.16" --> latest CI build of release-1.16
- "stable2" --> "latest-1.15" --> latest CI build of release-1.15
- "stable3" --> "latest-1.14" --> latest CI build of release-1.14

"beta" here will shifted back to be even with "dev" once we've deleted
the 1.14 jobs in a separate PR.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>