prepare_release_branch: Update generated config regex substitutions
prepare_release_branch.py uses a regex to replace the version markers of
a release branch during the beta.0 branch cut.

As we've moved to using the "latest**" version markers in
test_config.yaml, we need to update the regexes to properly recognize
the versions in the "k8sVersions" and "nodeK8sVersions" fields.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>