images: Symlink krte/variants.yaml to kubekins-e2e/variants.yaml
Recently discovering images/krte/variants.yaml, I noticed that this
variants.yaml closely mirrors the kubekins-e2e one, however it is not an
update that the Release Managers have made in the past, and it had some
stale entries. Instead of making checking the content between the two,
we symlink kubekins-e2e/variants.yaml so that they're always in sync.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>