Tolerate null commit values in job-flakes query
Some release-branch periodics were added but mistakenly dumped their
logs into pr-logs. The job-flakes query is unable to parse the commit
for these ci jobs masquerading as pull jobs, and so null ends up
propogating to commit_consistency. This then causes jq to error out
while attempting to convert null to a number

Add safeguards in the query and in the jq filter to prevent this
from breaking the query