Crier: Report periodics again
In https://github.com/kubernetes/test-infra/pull/15595 we changed
crier to respect the .Spec.Report field rather than handle it within
individual reporters.

This caused periodics to not get reported at all anymore, because their
.Spec.Report field always remains the zero value of a boolean, false.

This PR hardcodes the .Spec.Report field on periodics to true, because
the approach of "negate job.Reporter.SkipReport" used for Presubmits and
Postsubmits does not work for periodics, as they have no Reporter
embedded.