Update scheduler benchmark job to select tests to run.
We are in the progress of changing the way of running scheduler benchmark tests. As part of this, we create a new set of tests with different test prefixes. However, we do not want to run the new ones just yet until we have everything working e2e. 
This change makes sure the benchrmark test jobs only run the existing tests.