Add kind ipv6 conformance presubmit jobs
Serial jobs take more tha 1.30 hours to run and
are not ideal to run as presubmit because of that.
However, we need signal on them to detect early
possible breakages in conformance.
We add one presubmit job whithout serial tests that
run on changes on the "test" folder.
We add an optional presubmit job that run ALL
conformance tests.