Handle 404 from GitHub API when assigning issues
The AssignIssue method from GitHub client now handles a 404 response from
GitHub API by returning all specified users as invalid, because a 404
response does not indicate the invalid user(s).

This is needed by the assign plugin to avoid silently failing if any of
the specified users is not valid.

Signed-off-by: Arash Deshmeh <adeshmeh@ca.ibm.com>