Kops - Switch from beta/stable version markers and use version-specific markers
Removing one of the jobs because we already have complete coverage. 1-18 job tests the latest 1.18 commit and the `kops-aws` job tests the latest master branch (which will diverge when k8s release-1.18 branch is cut)
I also renamed the clusters to not use a period in order to avoid any potential subdomain issues if we choose to test ACM certificates down the road.