Try capping bazel's memory in some pull-kubernetes-e2e jobs
My working theory is bazel's jvm is asking for more memory than is
available, and getting oomkilled. Instead, let's cap its memory so
that it tries to GC instead of grow memory.

The 6Gi number was there for -gce before, capping bazel's jvm to 4Gi
is a total guess to leave some overhead for non-bazel stuff.

Add a 6Gi resource request for -rbe so we can compare apples and oranges
to see how this affects RBE and non-RBE.