kops: Ensure kops installs the same version of k8s we're using
Some tests are tightly bound to a particular version of k8s, for
example https://github.com/kubernetes/kubernetes/pull/87440 both
changed the behaviour and the tests in one PR.  We need therefore to
run the exact same k8s version as the tests.  This is a follow-on to
k8s version we download.