Boskos: remove global lock
Boskos always uses a CRD storage, which means that we get the
built-in optimistic concurrency from Kubernetes. This allows us
to get rid of the lock, as the Update requests we do will fail in
case another request was quicker and updated the object in the meantime.

From a users POV this means that there is a higher chance of
Acquire/AcquireByState requests failing, but a much faster response
time.

As a follow-up, we could make the Acquire/AcquireByState handlers retry
the whole allocation to reduce the likelyhood of transient failures
there.