Update Windows log collector for aks-engine
Previously it was assumed that there would be a corresponding
SSH private key for the provided SSH public key (ssh_key.pub and
corresponding ssh_key). This turns out to not always be the case.

This patch introduces a dedicated parameter for the ssh private key
as to not be dependant on this specific assumption.