Run improved scheduler performance benchmark in a new JJ
Recently, we have introduced a new test suite which in addition to golang benchmarking
collects metrics. In order to be able to compare the new test suite with currently existing one (ci-benchmark-scheduler-master job),
we would like to run both suites in parallel (e.g. for the next few releases).
Once we are confident about the new test suite, we can depricated the current one and drop it's job.