Kops - try computeMetadata for external IP before using ifconfig.co
This logic was removed from kops-e2e-runner.sh because it wasnt falling back properly after the computeMetadata requests would fail due to workload identity being enabled on the cluster.
Now that we can use a ServiceAccount on our prow jobs that will enable access to the computeMetadata URLs, we can try to use it first before ifconfig.co.

We've also been getting occasionally blocked from ifconfig.co, presumably due to high volume, so this will help with test flakes.

The prow jobs are not yet configured with a service account but this logic can be added independently of that.