Switch --kops-overrides field to string
With pflag.StringSlice, the flag wasn't actually getting registered (check `kubetest --help` to confirm).

I believe this is because in main.go we create a new flag set and parse them [0] from the defineFlags() function which doesnt use any pflag defintions from `init()`

Rather than splitting up the kops flag definitions so that the one pflag StringSlice can be defined properly I figured it was easier to just switch it back to flag.String to be consistent with the others.
We end up treating the values as a long string passed to `kops create cluster`'s single `--override` flag anyways.

[0] https://github.com/kubernetes/test-infra/blob/da16e1b1780da64f9802c3538ccec06332e259e2/kubetest/main.go#L279-L284