github: use a cacheable call when ghproxy is configured
The V3 API call to check if one specific user is a collaborator on a
repository does not currently have support for e-tags, which makes this
call not cache-able by ghproxy. Therefore, this call has an outsize
impact on the token conspumption of a Prow instance. If we can detect
that our GitHub client is pointed at a ghproxy instance (using a
best-effort approach) we can choose to use a more expensive but
cacheable call instead and thereby reduce the token usage.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>