Fix broken CI job - ci-kubernetes-e2e-gci-gce-ipvs
The CI job fails as the 3 worker nodes fail to come up. A quick look at
the kubelet logs show:
```
F0318 21:59:03.419573    3167 server.go:274] failed to run Kubelet: docker API version is older than 1.26.0
```
So essentially the image has a stale runtime.

We need to resurrect the broken CI job. the closest working CI job to
this one is "ci-containerd-e2e-ubuntu-gce". With the updates the
configure.sh from the `containerd/cri` updates the runtime before
installing kubernetes.

Signed-off-by: Davanum Srinivas <davanum@gmail.com>