ghproxy: handle more paths
It seems like the GitHub V3 REST API allows for access to large swaths
of the API under different roots. For instance, a user can declare a
repository with `/repos/:org/:repo` using the plain-text names of these
concepts, or with `/repositories/:repoId` using the numeric identifier.
This patch changes our path handler to allow for path starting with
either approach for repositories and organizations.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>