Kops - increase job timeout for updown e2e job
This has been failing recently and the job timeouts before the log artifacts get dumped.
This is because it takes a long time to finally give up on validating that the cluster is healthy (each attempt takes 2m30s).

I can work on shortening that timeout but to first get some artifacts i'm increasing the job timeout which requires increasing the interval as well.