bugzilla: announce PR states when evaluating post-merge
Previously, when we handled a PR merging we'd check the linked bug for
PRs and move it if all of them were merged, but we were light on details
for which PRs we'd checked and we posted nothing if there was nothing to
do. This change updates our messaging to be more explicit and posts an
explanation fo why we did nothing, when we do nothing.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>