releng: Loosen blockade regex on kubernetes/release repo
Since the blockade was initiated, we've migrated several of the
blockaded tools over to go and added tests for them.

This also fixes a bug where the blockade was being triggered
unintentionally in subdirectories (like 'images/build/cross') because
the regex was not properly grouped ('^foo|bar|baz' as opposed to
'^(foo|bar|baz)'.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>