kops grid: start to populate the grid
We don't have a lot of data to go on at the moment when it comes to
CNI providers.  flannel is exhibiting some problems (at least with
webhooks, possibly more, on rhel family distros), but we don't really
know whether that is specific to flannel.

Start to fill out the grid with a few more networking plugins, mostly
chosen as the starting set because they are the ones I am most
familiar with.

We're starting off running these once per day by default, we'll drop
it to weekly after a few days.