Add k8s-infra-prow-build's boskos to monitoring
I don't have access to the prometheus instance for
monitoring.prow.k8s.io so I'm not sure what the existing metrics look
like. This should work well with the existing boskos resources
dashboard.

I'm not sure how well this will work with the boskos server
dashboard. Do the scraped metrics automatically have a `host` or `target`
tag included in them that could be added as a parameter?