Modify gci option in --extract flag
By default, the --extract=gci/FAMILY flag will look at COS images in
the GCP project "container-vm-image-staging" and use the GCS bucket
"gs://container-vm-image-staging/k8s-version-map" to map the image name
to k8s version to be extracted. These values are hard-coded, and it
would be ideal to pass that information to the --extract flag instead.