Stop publishing latest-green.txt and migrate to k8s-infra-prow-build
This migrates ci-kubernetes-e2e-gci-gce over to the k8s-infra owned
prow build cluster. The release branch variants of this job have
already been moved over. What was preventing this from moving over
was the --publish flag, implying that this job needed write access
to a bucket that won't allow non-google.com access.

I have looked using cs.k8s.io and can't find any code that actually
relies on the existence of latest-green.txt. I could have sworn that
it was at one time used by tooling in kubernetes/release (eg: as part
of find_green_build), but I haven't found it in history there either.

I've looked in this repo's history, and publishing to this path has
been present since inception, but I still can't find what might actually
be looking for this file.

It's referenced in docs in a few places, which I will PR to remove if
those are the only references.