Update boskos pool dashboard to support same type, different instance
I would like to have pools named identically across prow-builds' boksos
and k8s-infra's boskos, so there's one less thing to change when
migrating a prowjob from one build cluster to the other.

k8s-infra-gce-project vs. gce-project was a mistake I'd like to undo

I also want to be able to display a panel per-instance for the same pool
type. This is the dumb hardcoded approach