image pushing: more Kubernetes-CSI repos and branches, separate dashboard
As discussed in https://github.com/kubernetes/test-infra/pull/17741
it's fine to enable jobs that will not succeed.

A separate dashboard gets added because it's easier than trying to
determine per repo where the job is meant to be shown and because it
might be useful to check all of those jobs at one glance.