Tag KUBE_CONTAINER_RUNTIME=docker explicitly
If we ever have to switch default runtimes in kube-up scripts, we need
to mark jobs explicitly. Let us start with the presubmit for sig-node
that run the node e2e tests. These tests depend on docker explicitly.
Note that there are specific containerd jobs that have a preset for
containerd as well if they explicitly need containerd.

Signed-off-by: Davanum Srinivas <davanum@gmail.com>