Crier: Retry when PJ update after report failed
Currently we never retry a prowjob if we reported successfully but
failed to update the prowjob status. This was presumably done with the
thought that we don't want to report a second time.

However, the net result is that we will still report a second time when
either the prowjob gets modified or crier is restarted but its a lot
harder to correlate the report to the previous error.