Refactored common functionality and added basic boskos support
Boskos project acquisition is still not as sophisticated as
in kubetest, but this provides a baseline to build on.

Moved all init logic to a common location. Eventually this init will
also encapsulate loading a struct stored on disk and enforcing
run only once.

Though still needs to be given to running up and down separately,
which will certainly require boskos information to be stored on disk.
Will also have to investigate how long boskos resource timeouts are.
If the struct is loaded after that timeout for down, we will have to
error because we can't guarantee access to the resource.

Simplified IsUp() checking because it currently needs a limited amount
of information to succeed.