Allow /approval variant for Prow approve plugin.
It's common within internal Google code review to see "LGTM, Approval"
as a response to approved code reviews, so I've been somewhat conditioned to use
`/approval`, only to discover later that it didn't work.

This adds the ability for Prow to accept both `/approve` and `/approval`
as valid commands.

This also makes it easy to add other variants (e.g. `/approved`?) later
if we need to.