releng: Drop test config for beta jobs, add stable4 jobs
beta markers shift depending on the stage in the release cycle and can
end up being inaccurate without consistent care and feeding of the
generated jobs.

Here we remove the beta job configs (not the k8s-beta version marker;
that'll happen in an upcoming commit) and establish a new 'stable4'
version config.

This serves a few purposes:
- beta releases are now cut from the main ('master') branch, so the
  'dev' 'k8sVersions'/'nodeK8sVersions' configs are now sufficient to
  describe the beta phase of the release cycle
- now that we've expanded the support cycle for Kubernetes versions
  (from 1.19 on) to 12+ months, removing the previous jobs no longer
  falls neatly in line with the timing of the next branch cut. The
  'stable4' job configs will now allow us to maintain generated jobs
  through the newly extended support cycle.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>