Address a auth race condition.
Also add --token-path for easier local development (no need to insert this token into a cookie).

When gerrit starts up it often sees transient unauthorized errors.
This is likely because we add authentication to the handlers in a go routine.
This change makes it set up all the handlers before forking a routine to keep
the auth up to date.

Also remove some pointless functions.