sig-cli: Disable forking for kubectl-skew jobs
kubectl skews jobs consistently cause config forking issues because of
the way they use generic version markers and oddities in the job name.

The kubectl skew jobs are also defined within a single file, instead of
alongside the other release branch jobs, which means they don't take
advantage of forking and are frequently stale.

We disable forking for the main kubectl-skew job to unblock creation of
the 1.19 branch release jobs.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>