Use test-cmd-args alongside test_args in gce-windows jobs
Switching from test-cmd-args to test_args in #18372 successfully
disabled the tests that were causing the jobs to fail. However, it also
enabled a whole set of new failing jobs that fit the ginkgo.skip
pattern. This provides both flags for the time being until a solution to
consolidating this behavior to one flag or the other can be implemented.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>