Removed unnecessary label from kt2 canary job
This label is being removed because, while it fixes an issue that
was occurring, it does more than it needs to and obscures intent.
The preset-k8s-ssh label fixes an issue with the GCE deployer's
log dump script by setting USER to "prow". It also does extra,
unnecessary work with SSH keys and does not resolve the underlying
issue. We have opted to fix this issue in the deployer itself by
adding logic around the USER env var.