Adds preset-windows-repo-list-2004 label to 2004 job
Those jobs will be running the test "should be able to pull from private registry with secret", and
they will pull the image from a private registry they need the credentials for.

Those jobs will pull a different image-repo-list file to use. We can't do this for every job, because
the PR that allows the tests to use a custom docker config.json file merged in 1.19.