Add cache support to prow
This commits adds basic caching support to prow by the usage of the
already available Object Storage bucket credentials.

Caching works only if job decorations are enabled, because it will hook
into the already existing `clonerefs` init container as well as the
`sidecar`. This means that we try to copy the cache from the Object
Storage bucket **after cloning the sources** and try to upload the cache
**after the main jobs process has finished successfully**.

We use `emptyDir` volumes to share the data during the execution, which
means that we create a new Volume for each `Cache` entry and mount them
into the `clonerefs`, `sidecar` and possible job containers.

A single cache entry can be defined in the `decoration_config` via:

```yaml
decoration_config:
  caches:
  - path: /tmp/gopath
    keys:
    - go.sum
    - version-marker.txt
    download_only: true
```

One caveat right now is that the `path` has to be unique because it will
be used as normalized reference for the volume and mounts. Another point
worth mentioning is that the `path` has to be absolute to be mount-able.

If `download_only` is true, then it will not upload the cache if the test
process succeeded and no cache is already available.

All `Cache.Keys` will be translated into a uniquely identifiable
tarball (see `cache.TarBallName()`), which stores the cache data and
therefore indirectly invalidates the cache if any of the `keys` change.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>