Add periodic conformance test run against cloud-provider-gcp
Adds a new periodic job to run the conformance ginkgo suite against
the cloud-provider-gcp repository. Test job is based on
ci-kubernetes-gce-conformance-latest, with the main difference being
using the new kubetest2 GCE deployer instead of kubetest. Because
cloud-provider-gcp does not have published release tars, this job
has to clone and build the cloud-provider-gcp repository. The job
also does not include the forking and the alerting in the job it is
based on because it needs to be validated in production prow before
meaningful decisions can be made using it.